Moved Onvista PDF tests to separate package
