+++ name.abuchen.portfolio.tests/src/name/abuchen/portfolio/datatransfer/pdf/onvista/OnvistaErtragsgutschriftErträgnisgutschrift.txt
